Improve name generation
Generating suitable names for generated types is a complex task, and the
current implementation has some shortcomings. This commit improves the
existing `Naming` trait and its implementations to make it easier to use.
