+++ xsd-parser/tests/feature/mod.rs
@@ -61,6 +61,7 @@ mod simple_type;
 mod simple_type_emptiable_complex_base;
 mod simple_type_max_length;
 mod static_list;
+mod trailing_underscore_group;
 mod tuple_with_integer;
 mod tuple_with_string;
 mod tuple_with_vec;
