+++ xsd-parser/tests/feature/trailing_underscore_group/schema.xsd
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Test for issue #197: Duplicate names when types differ only by trailing underscore -->
+<xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema"
+            xmlns="http://example.org/test"
+            targetNamespace="http://example.org/test"
+            elementFormDefault="qualified">
+
+  <xsd:group name="MyGroup">
+    <xsd:sequence>
+      <xsd:element name="Name" type="xsd:string" minOccurs="0"/>
+    </xsd:sequence>
+  </xsd:group>
+
+  <xsd:complexType name="Base_">
+    <xsd:sequence>
+      <xsd:group ref="MyGroup"/>
+    </xsd:sequence>
+  </xsd:complexType>
+
+  <xsd:complexType name="Base">
+    <xsd:complexContent>
+      <xsd:restriction base="Base_">
+        <xsd:sequence>
+          <xsd:group ref="MyGroup"/>
+        </xsd:sequence>
+      </xsd:restriction>
+    </xsd:complexContent>
+  </xsd:complexType>
+
+</xsd:schema>
